Refactor: model 폴더 생성, validate 추가
