+++ src/js/model/model.js
@@ -1,0 +1,13 @@
+export class Car {
+  constructor(name) {
+    this.name = name;
+    this.position = 0;
+  }
+
+  forward() {
+    this.position += 1;
+  }
+}
+
+export const getCarClassList = (element) =>
+  element.value.split(",").map((carName) => new Car(carName.trim()));
